% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractors.R
\name{contractors}
\alias{contractors}
\title{Повертає довідник типів контрагентів}
\usage{
contractors()
}
\value{
Датафрейм із довідником типів контрагентів
}
\description{
Повертає довідник типів контрагентів
}
