% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{download_organisations}
\alias{download_organisations}
\title{Download Ogranizations.}
\usage{
download_organisations(filename)
}
\arguments{
\item{filename}{Full path to file where you want organisations to be saved.}
}
\description{
Function to download information on all registered organizations on website.
}
\examples{
download_organisations()
}
\keyword{download_organisations}
