% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{get_organizations}
\alias{get_organizations}
\title{Повертає перелік всіх розпорядників, зареєстрованих на spending.gov.ua}
\usage{
get_organizations()
}
\value{
Датафрейм із переліком всіх розпорядників бюджетних коштів, зареєстрованих на порталі spending.gov.ua
}
\description{
Повертає перелік всіх розпорядників, зареєстрованих на spending.gov.ua
}
