% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{last_date}
\alias{last_date}
\title{Get the date of the latest transaction}
\usage{
last_date()
}
\description{
Function to load the date of the latest updates in spending.gov.ua database
}
\examples{
last_date()
}
\keyword{last_date}
