% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastload.R
\name{lastload}
\alias{lastload}
\title{Повертає дату останнього завантаження даних до БД API транзакцій}
\usage{
lastload()
}
\value{
Датафрейм, який містить дати останнього завантаження даних до БД API
}
\description{
Повертає дату останнього завантаження даних до БД API транзакцій
}
