% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{orgs}
\alias{orgs}
\title{Get organizations ids}
\usage{
orgs(organizations)
}
\arguments{
\item{organizations}{Character vector of organizations names or the beginnings of their names.}
}
\description{
Function to get organizations codes ("edrpous") given their names (or the beginnings of their names)
}
\examples{
orgs()
}
\keyword{organizations,}
\keyword{orgs}
