% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{orgs_ids}
\alias{orgs_ids}
\title{Get organizations ids}
\usage{
orgs_ids(organizations, filename = "organizations.csv")
}
\arguments{
\item{organizations}{Character vector of organizations names or the beginnings of their names.}

\item{filename}{The path to file with registered organizations. Default - "organizations.csv"}
}
\description{
Function to get organizations codes ("edrpous") given their names (or the beginnings of their names)
}
\examples{
orgs_ids()
}
\keyword{org_ids}
\keyword{organizations,}
