% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping_transactions.R
\name{ping_transactions}
\alias{ping_transactions}
\title{Перевірка доступності API транзакцій}
\usage{
ping_transactions()
}
\value{
Повертає код відповіді сервера
}
\description{
Перевірка доступності API транзакцій
}
