% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{regions}
\alias{regions}
\title{Повертає список регіонів із кодами}
\usage{
regions()
}
\value{
Датафрейм, який містить перелік регіонів із їхніми кодами. Ці коди можна використовувати у запитах, що стосуються транзакцій, а також під час обробки таблиць із транзакціями
}
\description{
Повертає список регіонів із кодами
}
