% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{regions_ids}
\alias{regions_ids}
\title{Get regions ids}
\usage{
regions_ids(regions = NULL)
}
\arguments{
\item{regions}{Character vector of regions names or the beginnings of their names. If not specified, the function returns named vector with all regions' ids.}
}
\description{
Function to get regions' ids given their names (or the beginnings of their names)
}
\examples{
regions_ids()
}
\keyword{regions,}
\keyword{regions_ids}
