% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{regs}
\alias{regs}
\title{Get regions ids}
\usage{
regs(regions = NULL, fixed = TRUE, ignore.case = TRUE)
}
\arguments{
\item{regions}{Character vector of regions names or the beginnings of their names. If not specified, the function returns named vector with all regions' ids.}

\item{fixed}{Logical. If TRUE, pattern is a string to be matched as is. If FALSE, regular expression is used.}

\item{ignore.case}{if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.}
}
\description{
Function to get regions' ids given their names (or the beginnings of their names)
}
\examples{
regs
}
\keyword{regions,}
\keyword{regs}
