% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{top100}
\alias{top100}
\title{Get 100 biggest transactions}
\usage{
top100(regions = NULL)
}
\arguments{
\item{regions}{Integer vector of regions ids. If not present, the function will return 100 biggest transactions in every region and on the national level.}
}
\description{
Function to load 100 biggest transactions within regions or anywhere.
}
\examples{
top100()
}
\keyword{top100}
