% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_transactions.R
\name{top_transactions}
\alias{top_transactions}
\title{Повертає сто найбільших транзакцій}
\usage{
top_transactions(region = NULL)
}
\arguments{
\item{region}{Код регіону. Можна отримати за допомоги функції regions(). Якщо параметр не зазначено, використовується параметр за замовчуванням ('Вся Україна')}
}
\value{
Датафрейм, який містить дані про сто найбільших транзакцій
}
\description{
Повертає сто найбільших транзакцій
}
