% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{transactions}
\alias{transactions}
\title{Отримайте дані транзакцій}
\usage{
transactions(payer = NULL, recipt = NULL, start = NULL, end = NULL,
  region = NULL, save = FALSE)
}
\arguments{
\item{payer}{Код ЄДРПОУ платника. Не може містити менше восьми символів}

\item{recipt}{Код ЄДРПОУ отримувача. Не може містити менше восьми символів}

\item{start}{Дата початку у форматі YYYY-MM-DD}

\item{end}{Дата закінчення у форматі YYYY-MM-DD}

\item{region}{Числовий код регіону, можна отримати за допомоги функції regions()}

\item{save}{Логічне (TRUE/FALSE), записувати результат пошуку у CSV файл (TRUE), чи повертати звичайний датафрейм (FALSE). Значення за замовчуванням - FALSE}
}
\value{
Датафрейм із даними транзакцій
}
\description{
Отримайте дані транзакцій
}
