% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{download_organisations}
\alias{download_organisations}
\title{Download Ogranizations.}
\usage{
download_organisations(folder = ".")
}
\arguments{
\item{folder}{Folder, where you want to save .csv file with ogranizations. Defaults to ".".}
}
\description{
Function to download information on all registered organizations on website.
}
\examples{
download_organisations()
}
\keyword{download_organisations}
