% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{orgs}
\alias{orgs}
\title{Get organizations ids}
\usage{
orgs(organizations, fixed = TRUE, ignore.case = TRUE)
}
\arguments{
\item{organizations}{Character vector of organizations names or the beginnings of their names.}

\item{fixed}{Logical. If TRUE, pattern is a string to be matched as is. If FALSE, regular expression is used.}

\item{ignore.case}{if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.}
}
\description{
Function to get organizations codes ("edrpous") given their names (or the beginnings of their names)
}
\examples{
orgs()
}
\keyword{organizations,}
\keyword{orgs}
