% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{transactions}
\alias{transactions}
\title{Get transactions}
\usage{
transactions(payers_edrpous = NULL, recievers_edrpous = NULL,
  regions = NULL, startdate = NULL, enddate = NULL, source = NULL)
}
\arguments{
\item{payers_edrpous}{Character vector of payers' codes ("edrpous") in looked transactions.}

\item{recievers_edrpous}{Character vector of receivers' codes ("edrpous") in looked transactions}

\item{regions}{Integer vector of regions ids. If not present, the function will return transactions in every region and on the national level.}

\item{startdate}{The first date of wanted period. Format - "yyyy-mm-dd"}

\item{enddate}{The last date of wanted period. Format - "yyyy-mm-dd"}

\item{source}{The source of transactions. At the moment (June 2018) there are two possible values: "DKSU" (Treasury) or "UZ" (Railway)}
}
\description{
Function to load all transactions, limited by payers' and / or receivers' codes or within sinlge day
}
\examples{
transactions()
}
\keyword{transactions}
