% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e-data.R
\name{transactions}
\alias{transactions}
\title{Get 100 biggest transactions}
\usage{
transactions(payers_edrpous = NULL, recipt_edrpous = NULL, regions = NULL,
  startdate = NULL, enddate = NULL)
}
\arguments{
\item{payers_edrpous}{Character vector of payers' codes ("edrpous") in looked transactions.}

\item{recipt_edrpous}{Character vector of receivers' codes ("edrpous") in looked transactions}

\item{regions}{Integer vector of regions ids. If not present, the function will return transactions in every region and on the national level.}

\item{startdate}{The first date of wanted period. Format - "yyyy-mm-dd"}

\item{enddate}{The last date of wanted period. Format - "yyyy-mm-dd"}
}
\description{
Function to load 100 biggest transactions within regions or anywhere.
}
\examples{
transactions()
}
\keyword{transactions}
